Refactor storybook code
