+++ .storybook/stories/playground/story-editor/api/media.js
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * Internal dependencies
+ */
+import { getDummyMedia } from '../getDummyMedia';
+
+export const getMedia = (params) => {
+  const dummyMedia = getDummyMedia();
+  const mediaResponse = {
+    data: dummyMedia,
+    headers: {
+      totalItems: dummyMedia.length,
+      totalPages: 1,
+    },
+  };
+
+  if (params.searchTerm) {
+    mediaResponse.data = dummyMedia.filter((media) => {
+      return media.alt.toLowerCase().includes(params.searchTerm.toLowerCase());
+    });
+    mediaResponse.headers.totalItems = mediaResponse.data.length;
+  }
+
+  return Promise.resolve(mediaResponse);
+};
